@@ -3,46 +3,77 @@
   "version": "1.0.0",
   "description": "",
   "main": "dist/backend/index.js",
+  "type": "module",
   "scripts": {
-    "test": "jest",
     "build": "parcel build",
-    "dev": "concurrently \"parcel watch\" \"nodemon dist/backend/index.js\"",
-    "start": "node dist/backend/index.js"
+    "dev": "concurrently \"parcel watch\" \"nodemon --env-file=.env dist/backend/index.js\"",
+    "start": "node --env-file=.env dist/backend/index.js",
+    "lint": "eslint .",
+    "test": "node --env-file=test.env --experimental-vm-modules --no-warnings node_modules/jest/bin/jest.js",
+    "test:watch": "node --experimental-vm-modules --no-warnings node_modules/jest/bin/jest.js --watch"
   },
   "targets": {
     "frontend": {
-      "source": "./view/index.html",
+      "source": "./src/view/index.html",
       "context": "browser"
     },
     "backend": {
-      "source": "./server/index.ts",
-      "context": "node"
+      "source": "./src/server/index.ts",
+      "context": "node",
+      "outputFormat": "esmodule",
+      "isLibrary": true
     }
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@testing-library/jest-dom": "^6.1.4",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.5.1",
     "@types/express": "^4.17.20",
+    "@types/jest": "^29.5.7",
     "@types/react": "^18.2.29",
     "@types/react-dom": "^18.2.14",
     "@types/supertest": "^2.0.15",
-    "@typescript-eslint/eslint-plugin": "^6.8.0",
-    "@typescript-eslint/parser": "^6.8.0",
+    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/parser": "^6.9.0",
     "concurrently": "^8.2.2",
-    "eslint": "^8.51.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-prettier": "^5.0.1",
+    "eslint": "^8.52.0",
+    "eslint-config-xo": "^0.43.1",
+    "eslint-config-xo-typescript": "^1.0.1",
     "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "nodemon": "^3.0.1",
     "parcel": "^2.10.0",
     "process": "^0.11.10",
     "supertest": "^6.3.3",
-    "ts-jest": "^29.1.1"
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.2.2"
   },
   "dependencies": {
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@hookform/resolvers": "^3.3.2",
+    "@mui/icons-material": "^5.14.18",
+    "@mui/material": "^5.14.15",
+    "@mui/x-data-grid": "^6.18.2",
+    "@mui/x-date-pickers": "^6.18.2",
+    "body-parser": "^1.20.2",
+    "dayjs": "^1.11.10",
     "express": "^4.18.2",
+    "express-validator": "^7.0.1",
+    "firebase": "^10.6.0",
+    "firebase-admin": "^11.11.0",
+    "firebaseui": "^6.1.0",
+    "mysql2": "^3.6.3",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.48.2",
+    "react-hot-toast": "^2.4.1",
+    "swr": "^2.2.4",
+    "wouter": "^2.12.1",
+    "zod": "^3.22.4"
   }
 }
